refactor(feedback): move types, schemas, and utils into dedicated folders
