+++ apps/dashboard/src/utils/feedback-page-url.ts
@@ -1,0 +1,7 @@
+export function getFeedbackPageUrl(pathname: string): string | undefined {
+  if (typeof window === "undefined") {
+    return;
+  }
+
+  return `${window.location.origin}${pathname}`;
+}
